fix friend function template declaration lookup
https://en.cppreference.com/w/cpp/language/friend
Note: A name first declared in a friend declaration within class or
class template X becomes a member of the innermost enclosing namespace
of X, but is not visible for lookup (except argument-dependent lookup
that considers X) unless a matching declaration at the namespace scope
is provided.

https://en.cppreference.com/w/cpp/language/adl
Note: Until C++20, although a function call can be resolved through ADL
even if ordinary lookup finds nothing, a function call to a function
template with explicitly-specified template arguments requires that
there is a declaration of the template found by ordinary lookup,
otherwise, it is a syntax error to encounter an unknown name followed
by a less-than character.